refactor: userapi moved to the required folder
