+++ src/pages/UsersPage/UsersPage.tsx
@@ -1,5 +1,5 @@
 import { useSelector } from "react-redux"
-import { useGetUsersQuery } from "../../entities/users/model/api/usersApi"
+import { useGetUsersQuery } from "../../entities/users/api/usersApi"
 import { withLoading } from "../../shared/lib/hoc/withLoading"
 import { UserList as OriginalUserList } from "../../widgets/UserList/UserList"
 import { selectAllUsers } from "../../entities/users/model/slice/userSlice"
